third_party/blink/web_tests/platform/generic/fast/dom/HTMLElement/class-list-gc-expected.txt

This tests that properties on the classList persists GC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS d.classList.life is 42
PASS d.classList.life is 42
PASS successfullyParsed is true

TEST COMPLETE
